Lots of improvements
1. Ensure resolution of fbo is > than screen.
This matches ofx Post Effects plugin meaning there is no resolution jump when activating effects.
I suspect this might make things sharper since we are working on a larger fbo for calculations that we scale back down to the screen. But maybe not, in which case this is a performance waste. Time will tell.

2. Stack FX’s
Now you can combine multiple fx’s. This is dangerous as activating multiple can kill the graphics card fast. Considering limiting only one of each active at any one time.

3. Control Bloom
Attach into fx passes with configurable params. This means we can change the brightness of the bloom which looks lovely.